Fix for bug##58750: 'out of connection objects' due to missing close() of prev. full_table_scan()'
If ha_ndbcluster::read_range_first_to_buf() had to do a ::full_table_scan() 
on a UNIQUE_INDEX, the resulting ScanOperation was not properly closed at end
of its lifetime. If lots of such operations is executed within the 
same transaction, we will eventually run out of operation or transaction objects.
(Transaction objects is due to a dumnmy Hupp'ed transaction object is allocated as
part of a ScanOperation)
      
This fix ensures that ::close_scan() is called for any open cursors at start of
ha_ndbcluster::read_range_first_to_buf().