BUG#59499: mtr check fails for some replication tests due to ipv6
By the time mtr processes the test file completely, the slave's
parameter MASTER_HOST exhibits a different value from the one
that was set when mtr had started processing the file. This was
caused by test cases that mess around with CHANGE MASTER TO
MASTER_HOST=... command and never restore the variable to its
original value (instead they restore it to a default value that
varies depending on ipv6 is available or not).

We fix this by saving (at the beginning) and restoring back (at
the end) the value for MASTER_HOST slave configuration parameter.