BUG#47557: mysqlbinlog -v should output 'limit 1' for update and            delete row events
mysqlbinlog -v shows pseudo-SQL statements, reconstructed out of
the rows, as comments. However, --verbose description mentions that
it reconstructs SQL statements, when in fact these are SQL-like
statements and not, 100% valid, SQL instructions.

We fix this by replacing "SQL" with "pseudo-SQL" in mysqlbinlog 
help text (pseudo-SQL is actually the term used in the user
manual).