Follow-up patch for test case for Bug #58816.
The test case turns off index condition pushdown as part of the test
but does not restore it to the value it had before the test case. This
makes all test cases following it run with index condition pushdown
turned off. The fix is to save the value of optimzier_switch before the
test starts and restore it after the test.