This is a fix for the Windows vs2008 compile problem that was introduced by rev 3423 (rb://563) of mysql-trunk-innodb.  The additional use of os_increment_counter_by_amount() and os_decrement_counter_by_amount() in ha_innodb.cc was rejected in vs2008 since the value being changed is a ulint while the Windows atomics function uses LONG or __int64 depending on whether the OS is 32 or 64 bit.  The type ulint is either unsigned __int64 or unsigned long int depending on the OS as well.   Since the definition of os_(in/de)crement_counter_by_amount did not have a typedef on the value being changed, it did not compile.
This same error was encountered by RB://538 at one point when similar Windows defines became inline for the first time.  

This change is a comprehensive fix for all windows atomic calls so that they are type casted correctly within a UNIV_INLINE function.  Type-casting in an inline function is better than typecasting in a macro because the input types are checked by the compiler.