Bug #58784  	rpl_row_ignorable_event fails on PB2
In RBR, The rows are inserted to a queue by the thread executing
the 'INSERT DELAYED' statement, and are taken out from the queue
by the handler thread to do the real insertion. Because these two
threads are running in parallel, there is a possibility that they
are run in an interleaved manner, and result in different number
of table_map and rows events.
      
Added 'wait' and 'signal' mechanism for the test to make the
binlog of multi 'INSERT DELAYED ...' stmt stable by forcing
every value is executed into one execution series, and then
each value will be binlogged into a separate rows event with
its table map event.