BUG#47853: Implicitly grouped queries do not need to use            temporary tables
A query is implicitly grouped if it contains an aggregate 
function but no GROUP BY clause. Implicitly grouped queries 
return zero or one row, so ordering does not make sense. 

This patch removes the order by clause during the prepare()
stage of the optimizer when implicitly ordered queries are 
detected.