Fix for Bug#58838 Wrong results with HAVING + LIMIT without GROUP BY                   when ICP is enabled.
The wrong result was caused due to "loosing" the HAVING condition
when changing to use a different index than originally selected by the
optimizer. 

During the optimization phase it was decided to use a secondary index
for retrieving data from the table. Since ICP was enabled the WHERE
condition was pushed down to the storage engine. A copy of the WHERE
condition was kept in the JOINTAB's pre_idx_push_select_cond
member. In the start of the execution phase it was determined that the
HAVING condition could be included in the JOINTAB's select
condition. Then in test_if_skip_sort_order() we (since part of the
select condition has been pushed down to the storage engine) replace
the JOINTAB's select condition with the original select condition
(found in the pre_idx_push_select_cond member). Since
test_if_skip_sort_order() determines that is will be less costly to
use the primary key for doing the retrieval of data (due to LIMIT) it
changes to use the primary key. If it had not changed index then
test_if_skip_sort_order() would have restored the select condition to
what it was when it was called. But due to changing index the part of
the select condition that has been pushed down must also be included
in the select condition. To achieve this the condition in
pre_idx_push_select_cond is kept as the JOINTAB's select
condition. This does not include the HAVING condition and as a result
the HAVING condition is no longer part of the JOINTAB's select
condition.

The fix for this problem: When adding the HAVING condition to the
JOINTAB's select condition also add it to the pre_idx_push_select_cond
(if this exists).