Bug#59799 handler statistics for the performance schema engine
Before this fix, the performance schema storage engine did not
update the handler_* statistics, so that:
  show status like "%handler%";
returned under evaluated values.

With this fix, calls to rnd_next(), rnd_pos(), etc
methods in the performance schema engine now also properly
update the corresponding handler_* statistics.

Tested manually.
No automated scripts provided, as the value reported for
a session in these counters is hard to predict exactly,
as it depends too much on the environment.