Bug #58709 assert in mysql_execute_command
The triggered assert checks that either an error has been reported or
the statement has been killed if execution of a stored routine fails.
It was triggered if a stored routine had a DELETE IGNORE statement
which failed but due to IGNORE had not reported any error.

This patch fixes the problem by changing mysql_delete() to return
a value indicating failure only if an error has been reported or
the statement has been killed.

Test case added to delete.test.