Bug #58455 Starting mysqld with defaults file without extension cause segmentation fault
Bug occurs because fn_expand calls fn_format
with NULL as ext.

This is a backport of the patch from 5.6.

Patch solve this problem by using an empty
string as extension, and adding assertions
to fn_format that correct arguments are passed.

It also add a test tests several variations of
using non-existing defaults files.