BUG#59708: 'CREATE USER ... IDENTIFIED WITH ..' statement is            binlogged wrongly
The server rewrites the CREATE USER statement for avoid logging
clear text passwords into the binary log when these are set
through IDENTIFIED BY clause. However, after WL#1054, there is
also the IDENTIFIED WITH clause, which is left unattended during
this rewrite procedure. The ultimate consequence is that those
clauses are discarded and not included in the statement that ends
up in the binary log.
      
We fix this by taking into consideration the IDENTIFIED WITH
clauses (along with its AS '...' suffix) when rewriting the
CREATE USER statement.