Bug#11763382: Assertion 'inited=INDEX in sql/handler.h
Used to be bug#56080.

The problem occurs because some TABLE objects are associated with
the optimizer execution structures even after close_thread_tables()
has been called. In turn, this means that final cleanup of some
TABLE objects is done from within THD::cleanup_after_query().
At this point in time, the TABLE objects may already have been
given to another session, meaning that we may end another session's
index or table scan.

The safest solution seems to be to always perform a thorough cleanup
of all execution data structures, including releasing the TABLE
objects, before close_thread_tables() is called. In other words,
this will have to be done at the end of each execution.

However, this is currently not possible in the implementation of
subquery materialization. This part of the optimizer assigns some
objects, including a TABLE object, for the lifetime of the
statement object, and not for each execution. This gives an odd
situation, where some parts of the data structures are set up for
reuse and others are not. Combine this with the fact that the
user may decide to use another strategy for a subsequent
subquery execution, the "permanent" data structures may not even
be used later.

It is therefore reasonable to allocate materialization execution
structures for the lifetime of one execution. It means that the
functions init_permanent() and init_runtime() of class
subselect_hash_sj_engine are combined into one setup() function.
In addition, the function Item_subselect::cleanup() will be called
when the cleanup() function for the underlying query expression
is called, and this function will cleanup and destroy the
materialization data structures, so that the subquery item object
is ready for a new optimization process in the next round of
execution.

However, the EXPLAIN functionality relies on these data
structures being present when explaining a query containing
a materialized subquery. Thus, we have two conflicting
requirements for query cleanup:
 - For maximum efficiency we should release resources as early as
   possible after execution.
 - We cannot release resources that are needed for EXPLAIN queries.

A partial solution to this problem is to split resource cleanup in two:
1. Delete JOIN objects as soon as possible - this makes sense 
   because TABLE objects are associated with them through JOIN_TAB.
   JOIN::destroy() is extended so that associations to joined tables
   are removed.
   But notice that JOIN objects may be needed for EXPLAIN queries,
   so sometimes deleting them must be delayed.
   Setting free_join= 0 inside mysql_select() when SELECT_DESCRIBE
   is part of select options is evidence of that.
2. Cleanup query expression objects (ie st_select_lex and
   st_select_lex_unit) later (in practice at end of query execution).
   One example: In mysql_explain_union() the call to unit->cleanup()
   was deleted, so now it is done by the general cleanup.

The bug fix does only a limited amount of refactoring in this area.

mysql-test/r/optimizer_debug_sync.result
  Results for new test case.

mysql-test/t/optimizer_debug_sync.test
  Test case for bug. Requires the debug sync facility, and often fails
  without the bug fix.

mysql-test/r/union.result
  Change in EXPLAIN EXTENDED output that shows an optimized predicate
  in the two query specifications within a UNION in a subquery.
  AFAIK, this is similar to how a non-subquery is reported, so the
  change is an improvement.

sql/ha_partition.cc
  Comment change.

sql/item_subselect.cc
  In Item_subselect::cleanup(), "new" engine is always deleted after
  an execution.
  Item_in_subselect::setup_engine() now builds the materialization
  engine for one execution only. Specific arena is no longer needed.
  In subselect_union_engine::cleanup(), call to
  unit->reinit_exec_mechanism is deleted because it is redundant.
  Implementation of cleanup() functions for subclasses of
  subselect_engine has been refactored.
  subselect_hash_sj_engine::setup() replaces init_permanent() and
  init_runtime().
  For class subselect_hash_sj_engine, freeing of the temporary
  result table is moved from the destructor to ::cleanup().
  Some changes necessary because of interface changes.

sql/item_subselect.h
  Some cleanup of the subselect_engine classes:
   - All virtual functions now marked virtual in all derived classes.
   - Functions that returned bool result are now declared accordingly.
   - Slight cleanup of private/protected/public performed.
   - Improved constructors.

sql/sql_lex.h
  Change in friend declaration.

sql/sql_select.cc
  Function JOIN::reinit() is renamed to JOIN::reset().
  JOIN::destroy() will now delete connection to associated tables.
  The st_select_lex_unit object is no longer cleaned up as early as
  before, but instead relying on the general cleanup of lex->unit
  in mysql_execute_command().

sql/sql_select.h
  A few small interface changes.

sql/sql_union.cc
  Cleaned up some confusing error handling in st_select_lex_unit::exec()
  and st_select_lex::cleanup().