Bug# 11833442 - GET RID OF THE NEED TO ACQUIRE TRX_SYS_T::LOCK IN S MODE IN LOCK_DEADLOCK_OCCURS
There is an unnecessary contention point in the deadlock checking where
we acquire the trx_sys_t::lock in S mode to reset the trx_t specific node
visited flag for deadlock checking. This should reduce the pressure on
trx_start and trx_commit a lot.

rb://610 Approved by Marko.