Bug#11765758  bug#58754 RPL_CORRUPTION FAILED on mysql-trunk
There are two outstanding issues: a crash and stopping slave with an unexpected error.

The current patch is hunting over the crash. Here is the first step to sort out
whether `+d' dbug instruction that leads to the error-log relates to the crash.
Notice, no real changes neither to the sources not the tests are done.
The idea is just to watch PB2 on the win 64.