Fix for Bug#11873324 "WRONG RESULT WITH ICP AND STRAIGHT_JOIN"
This bug causes NOT IN subqueries to in some cases return too few
rows. This can happen when we have a subquery where a table is
normally accessed using ref access and we have pushed an index condition
down to the storage engine based on this but in some cases the access
method is changed during execution to use table scan instead of ref
access. Queries like this are marked with 'Full scan on NULL key' in
the explain output.

Details about how this situation occurs.

1. Index condition pushdown: Since the optimizer decides that the
   access method for one of the tables in the subquery should be ref
   access this is candidate for using index condition pushdown. The ICP
   code will push part of the table's condition down to the storage engine.
   The pushed parts are then removed from the table's condition and will
   therefore not be evaluated by the server.

2. Next looking at what 'Full scan on NULL key' means for these queries:

This is a NOT IN subquery. So for every record produced by the outer
query we will execute the subquery and see if this particular value
is produced by the subquery. The important parts relevant for this
case is found in the following two functions:

A. longlong Item_in_optimizer::val_int() (called from
   evaluate_join_record via a few other Item::val... functions)

   This functions checks if the value to check for is NULL. If it is NULL then
   the function does (among other things):

    1. Turn off the predicates that are based on column compares for
       which the left part is currently NULL. This is done by setting
       the subquery's "cond guards" to false.

    2. Executes the subquery (see next step)

    3. Turn all predicates back on. This is done by setting the
       subquery's "cond guards" back to true.

B. The execution of the subquery will eventually call the following function:

    subselect_single_select_engine::exec()

   In this function we have the following relevant actions:

    1. We check if the subquery has "guarded conditions" and if so we do
       a detailed check for all join tabs of the subquery to see if any of
       the "guarded conditions" are "false" (see step 1 above).

    2. If any of the "guarded conditions" are false we do as commented in the
       code:

      /*
        For at least one of the pushed predicates the following is true:
        We should not apply optimizations based on the condition that was
        pushed down into the subquery. Those optimizations are ref[_or_null]
        acceses. Change them to be full table scans.
      */
 
       So we basically changes from using ref access to table scan.

Summary for how this produces missing results:

1. When optimizing the subquery we push an index condition down to InnoDB.
   This is allowed since ref access is chosen.

2. If the outer query produces a non-null value to be checked
   for "NOT IN subquery", the subquery is executed using ref access
   (and everything is fine).

3. If the outer query produces a NULL value the execution of the
   subquery is changed to use table scan. In this case the pushed index
   condition will no longer be evaluated by InnoDB and the subquery will
   produce more records than it should and the NOT IN will evaluate to false
   (when it should evaluate to true) and the query will produce fewer
   result records than it should (which is not fine).

The fix for this is to add a check in push_index_cond() to determine
if the jointab has "guarded conditions". If it has then do not push
any index conditions down to the storage engine.