bug#11748510 incident of deadlock on slave is overdramatized
In cases of temporary errors in replication event execution (deadlock,
timeout) the error log gained an overreacting error message whereas
just a warning would be fine.

Fixed with checking of the temporary status of the error inside
Slave_reporting_capability::report()  to demote the error to the
warning level if needed. 
The warning gets converted into the error
whenever number of successive attempts to re-apply the event(s) gets
equal to @@global.slave_trans_retries.