WL#5829 optimize log_flush_order_mutex usage in mtr_commit()
rb://620 approved by: Marko

During mtr_commit() we grab log_flush_order_mutex to ensure that we
can insert into the flush_list in the same order in which the mtrs
are committed. However, a page needs to be inserted in the flush
list iff it was clean before it is touched by the mtr. This
optimization is to acquire log_flush_order_mutex iff mtr has modified
any clean pages. If all pages modified by the mtr were already dirty
then they must already be in the flush list and therefore there won't
be any insertions made to the flush list