BUG#12324076 - RQG_RPL_SEMISYNC TEST FAILURE, SLAVE CRASHES WITH SEGMENTATION FAULT
The patch for BUG#11765887 BUG#58897 verifies if the configuration is not
crash-safe and prints out a warning message to alert users on that.

The routine that keeps track of warnings and errors on the slave was modified
in BUG#11748510 - BUG#36524 and after that report_function may be null if
global_system_variables.log_warnings is not set. However, report_function is
used without checking its value thus causing a crash.

Fixed the problem, by checking if report_function is null before using it.